# backend/openapi.yaml
openapi: 3.0.3

info:
  title: Todo API
  version: 1.0.0

paths:
  /todos:
    get:
      operationId: getTodos
      summary: Todo一覧を取得する
      responses:
        "200":
          description: Todo一覧の取得に成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
        "500":
          description: 予期せぬエラーが発生
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createTodo
      summary: Todoを作成する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
      responses:
        "201":
          description: Todoの作成に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 予期せぬエラーが発生
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /todos/{todoId}:
    get:
      operationId: getTodo
      summary: Todoを1件取得する
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Todoの取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          description: 該当のTodoが発見できない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 予期せぬエラーが発生
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTodo
      summary: Todoを更新する
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoRequest"
      responses:
        "200":
          description: Todoの更新に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          description: 該当のTodoが確認できない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 予期せぬエラーが発生
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteTodo
      summary: Todoを削除する
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Todoの削除に成功
        "404":
          description: 該当のTodoが確認できない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 予期せぬエラーが発生
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - status
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
          enum:
            - todo
            - done

    CreateTodoRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100

    UpdateTodoRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - todo
            - done

    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string